+++ dmitriev.arseny/I3/defCountDiffLett.cpp
@@ -6,11 +6,11 @@ size_t defCountDiffLett(const char* cStr)
 {
   size_t count = 0;
 
-  for (int i = 0, moreThanOne = 0; cStr[i]; i++, moreThanOne = 0)
+  for (size_t i = 0, moreThanOne = 0; cStr[i]; i++, moreThanOne = 0)
   {
     if (std::isalpha(cStr[i]))
     {
-      for (int j = i + 1; cStr[j]; j++)
+      for (size_t j = i + 1; cStr[j]; j++)
       {
         if (cStr[i] == cStr[j])
         {
